Enable dvdkit on solaris
solaris has iconv() in libc.so, a separate libiconv.so library does not
exist.  Only add a -liconv library when we were succesful compiling a
test program with the -liconv option.

--enable-debug does not work on solaris, because test "-g" in configure
script fails with a test error message "test: argument expected". Use
string comparision to test for a non empty $_debug / $_profile


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@7255 b3059339-0415-0410-9bf9-f77b7e298cf2