Allow character in the font description file to be in UTF-8.
A character defined in the font description file can be either
an ASCII character, any character in the range of 0x80 to 0xFF
or - to avoid character set problems, and that is recommended -
a character in UTF-8 encoding now.

Non-ASCII characters will be stored in the nonASCIIidx array.
The indices 0..127 of this array correspond to the indices
128..255 of the Fnt array.

(This also settles the "Translate messages shown in the GUI window(s)
from UTF-8" issue.)


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@32821 b3059339-0415-0410-9bf9-f77b7e298cf2