configure: disable internal libmpeg2 by default.
While the last tests showed that it can be slightly
faster in some cases, the difference should not be
relevant to the majority of users.
In addition, the code is of not particularly high
quality, so I consider it better/safer to have it
disabled by default.
Removing the code after the next release is a possible
future action.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@36755 b3059339-0415-0410-9bf9-f77b7e298cf2