configure: improve FFmpeg check.
If internal FFmpeg is not available we should fall back to
external automatically instead of trying to build without
(which is currently broken it seems).
Also we can compile without internal copy as long as the
necessary headers can be found.
Two are required that FFmpeg does not install:
libavformat/internal.h
libavutil/x86/asm.h


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@36878 b3059339-0415-0410-9bf9-f77b7e298cf2