demux_ty: Try to fix calculation of used file size.
TiVo files can contain junk at the end, and this code
finds where the "good" data ends.
Unfortunately there was not a shred of sense in the original
code.
The code now seems to do at least something sensible with the
few files I had to test.

Fixes a warning from Coverity, which also thought this
code made no sense.

git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@37806 b3059339-0415-0410-9bf9-f77b7e298cf2